refactor(dashmate): name new modules in kebab-case, per the package guideline
Review feedback, accepted over my earlier objection that the surrounding
directory is camelCase. The guideline covers the package, and files this change
introduces are the ones that can follow it without renaming code the change
does not otherwise touch.

Class files keep PascalCase. That is a separate and unambiguous convention here
- ConfigFile.js, Prescription.js, Samples.js - and the record's model and
repository are classes.
